custom_components/openwakeword-installer/requirements.txt
gitpython
homeassistant
voluptuous
